# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Meridian Integration Gateway
  version: 1.0.0
  summary: REST and webhook façade over Meridian Freight's AS/400 shipment feed and SOAP rate-quote service.
  description: |
    Design-first contract (ADR-P01-4). Shippers review this document before code exists, and every build
    runs conformance tests against it (Schemathesis), so every status code the service can return is listed here.

    **Errors.** Every error body is RFC 9457 Problem Details (`application/problem+json`). The `type` URI is
    stable and machine-readable; `title` is stable per `type`; `detail` is human-readable and may change.

    **Authentication.** Send your API key in `X-API-Key`. Keys are scoped per shipper: every read is filtered
    to your own records, and a record that belongs to someone else returns `404`, never `403`. Dead-letter
    endpoints additionally require a key with the `ops` scope; a shipper key gets `403`.

    **Upstream protection (ADR-P01-1).** Meridian's rate service accepts at most 5 concurrent requests. The
    gateway allows at most 4 in flight across all shippers. Beyond that, and while the circuit breaker is
    open, `POST /v1/rate-quotes` fails fast with `503` and a `Retry-After` header (seconds). Honour it, and
    add jitter to your own retries.

    **Idempotency.** `POST /v1/rate-quotes` requires an `Idempotency-Key`. Retrying with the same key and the
...
    returns `409` with `Retry-After`. Keys are retained for 24 hours.

...
  contact:
...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /v1/webhook-subscriptions:
    post:
      operationId: createWebhookSubscription
      tags: [Webhook subscriptions]
      summary: Register a webhook endpoint
      description: |
        The URL must be `https` and must resolve to a public address; private, loopback, link-local and
        metadata addresses are refused (`422`, `type` ending in `/webhook-url-not-allowed`), as are URLs
        with credentials, whitespace or control characters. A shipper can hold at most 25 subscriptions
        (`422`, `type` ending in `/subscription-limit-reached`). The response contains the signing `secret`
        exactly once (`Cache-Control: no-store`). Store it: it cannot be retrieved again.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WebhookSubscriptionRequest' }
      responses:
        '201':
          description: Subscription created. The signing secret is included this one time only.
          headers:
            Location:
              description: URL of the subscription.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookSubscriptionCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422':
          description: |
            The body failed validation (see `errors`), the URL is not allowed (`/webhook-url-not-allowed`;
            the detail never says why, so it cannot be used to probe our DNS), or the shipper already has
            25 subscriptions (`/subscription-limit-reached`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/webhook-subscriptions/{subscription_id}:
    parameters:
      - name: subscription_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getWebhookSubscription
      tags: [Webhook subscriptions]
      summary: Get a webhook subscription (never includes the secret)
      responses:
        '200':
          description: The subscription
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookSubscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteWebhookSubscription
      tags: [Webhook subscriptions]
      summary: Delete a webhook subscription
      description: |
        The subscription and its deliveries are removed: pending ones are never sent, and the delivery
        history goes with it. Webhook dead letters for it stay listed but can no longer be replayed (`422`).
      responses:
        '204':
          description: Deleted
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Per-shipper API key, stored server-side as a SHA-256 hash and rotated with overlap. Ops keys carry the
        `ops` scope (OpenAPI cannot express scopes for API keys, so it is documented on each ops operation).
  headers:
    RetryAfter:
      description: Seconds to wait before retrying (RFC 9110 delay-seconds form).
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    BadRequest:
      description: Malformed request (for example a cursor that is not valid or was issued for another filter).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or unknown API key.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The API key is valid but has the wrong scope (`ops` for dead letters, `shipper` for shipper endpoints).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: No such resource for this API key (records of other shippers also return 404).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PayloadTooLarge:
      description: Request body larger than 64 KB.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnsupportedMediaType:
      description: Request body must be `application/json`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationFailed:
      description: A parameter or the body failed validation; see `errors`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Per-key rate limit exceeded.
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalError:
      description: Unexpected error. The `instance` value identifies it in the gateway's logs.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details. Unknown extension members must be ignored by clients.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          description: Identifies the problem type, e.g. `https://errors.meridian-gateway.example/request-in-progress`.
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance:
          type: string
          format: uri-reference
          description: Identifies this occurrence (a request ID), for support tickets.
        errors:
          type: array
          description: Field-level validation errors (present on `422` validation failures).
          items:
            type: object
            required: [location, message]
            properties:
              location: { type: string, description: 'JSON pointer or parameter location, e.g. `/body/origin_zip`.' }
              message: { type: string }
...
    EventType:
      type: string
      enum: [shipment.created, shipment.status_changed, rate_quote.completed]
    WebhookSubscriptionRequest:
      type: object
      additionalProperties: false
      required: [url, event_types]
      properties:
        url:
          type: string
          format: uri
          pattern: '^https://'
          maxLength: 2048
        event_types:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/EventType' }
    WebhookSubscription:
      type: object
      required: [subscription_id, url, event_types, status, created_at]
      properties:
        subscription_id: { type: string, format: uuid }
        url: { type: string, format: uri }
        event_types:
          type: array
          items: { $ref: '#/components/schemas/EventType' }
        status:
          type: string
          enum: [active, disabled]
          description: '`disabled` after the endpoint answered `410 Gone`.'
        created_at: { type: string, format: date-time }
    WebhookSubscriptionCreated:
      allOf:
        - $ref: '#/components/schemas/WebhookSubscription'
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              pattern: '^whsec_[A-Za-z0-9+/]+={0,2}$'
              description: Standard Webhooks signing secret. Returned only in this response.
...